// src/data/types.ts
export interface FAQ {
  question: string;
  answer: string;
}